backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.10.3
langchain-core==0.3.27
langchain-community==0.3.13
langchain-google-genai==2.0.6
python-dotenv==1.0.1
google-generativeai==0.8.3